Uglify Card component
