--- README.md
@@ -1,1 +1,0 @@
-# Elfsight Test Task (Rick and Morty API)
