--- src/components/card/CardInfo.js
@@ -1,12 +1,0 @@
-import styled from 'styled-components';
-
-export function CardInfo({ children }) {
-  return <StyledCardInfo>{children}</StyledCardInfo>;
-}
-
-const StyledCardInfo = styled.div`
-  display: flex;
-  flex-direction: column;
-  color: #fff;
-  padding: 20px;
-`;
