--- src/components/card/index.js
@@ -1,3 +1,0 @@
-export { Card } from './Card';
-export { CardStatus } from './CardStatus';
-export { CardTitle } from './CardTitle';
